+++ packages/shared/src/index.ts
@@ -1,4 +1,7 @@
 export * from "./user/index.js";
 export * from "./zSchema/index.js";
-export * from "./types/index.js";
 export * from "./utils/index.js";
+
+export type * from "./types/index.js";
+
+export type { Prettify, ApiResponse } from "./types.js";
