+++ packages/shared/src/types.d.ts
@@ -1,14 +1,10 @@
-declare global {
-  interface ApiResponse<T = unknown> {
-    data?: T;
-    error?: unknown;
-    message: string;
-    success: boolean;
-  }
+export type Prettify<T> = {
+  [K in keyof T]: T[K];
+} & {};
 
-  type Prettify<T> = {
-    [K in keyof T]: T[K];
-  } & {};
+export interface ApiResponse<T = unknown> {
+  data?: T;
+  error?: unknown;
+  message: string;
+  success: boolean;
 }
-
-export {};
