+++ packages/shared/src/zSchema/health.ts
@@ -1,6 +1,5 @@
 import { time, timeStamp } from "console";
 import z from "zod";
-import { generateSuccessResponseSchema } from "../utils/index.js";
 
 export const checked = z.object({
   db: z.enum(["up", "down"]),
@@ -15,13 +14,8 @@ export const created = toCreate.extend({
   timeStamp: z.string(),
 });
 
-export const checkResponse = generateSuccessResponseSchema(checked);
-export const writeDeleteResponse = generateSuccessResponseSchema(created);
-
 export const health = {
   checked,
   toCreate,
   created,
-  checkResponse,
-  writeDeleteResponse,
 };
